pyproject.toml
[tool.poetry]
name = "parimate"
version = "0.1.0"
description = ""
authors = ["AI WorkShop Week"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "20.1"
pyexiftool = "^0.5.6"
pytest = "^8.3.5"
omegaconf = "^2.3.0"
urllib3 = "^2.3.0"
tensorflow = "^2.18.0"
deepface = "^0.0.93"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
